Add uv_resolution as testenv parameter
uv supports defining a custom resolution strategy, defaulting to
'highest', and allowing 'lowest' or 'lowest-direct' at the time of
writing, which provides an alternative to package constraints enabling
lowest direct or transitive dependencies to be installed.

This change provides a new 'uv_resolution' configuration option for test
environments, which if defined will apply the '--resolution' option to
uv to enable this behavior.

Implements #28.